Add files via upload
I needed to find a way I could get some data from a internal system without see or interage with It's code, so a automate the mouse and key board to make the movements that a user would do to get the data, copy to a notepad and print. this is the result.
The role script takes about 30seconds and the user can use that time to do something else.
The places with "****" would be able to identify the system so I suppress it to post.